# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puspa API Documentation",
        "description": "Dokumentasi API lengkap untuk layanan Puspa",
        "contact": {
            "email": "admin@puspa.sinus.ac.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/therapists": {
            "get": {
                "tags": [
                    "Therapists"
                ],
                "summary": "Mendapatkan daftar semua terapis",
                "operationId": "getTherapistsList",
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TherapistResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (bukan admin)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Therapists"
                ],
                "summary": "Membuat data terapis baru",
                "operationId": "storeTherapist",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TherapistCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Terapis berhasil dibuat"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/therapists/{therapistId}": {
            "get": {
                "tags": [
                    "Therapists"
                ],
                "summary": "Mendapatkan detail satu terapis",
                "operationId": "getTherapistById",
                "parameters": [
                    {
                        "name": "therapistId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TherapistResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data terapis tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Therapists"
                ],
                "summary": "Memperbarui data terapis",
                "operationId": "updateTherapist",
                "parameters": [
                    {
                        "name": "therapistId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TherapistUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update berhasil"
                    },
                    "404": {
                        "description": "Data terapis tidak ditemukan"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Therapists"
                ],
                "summary": "Menghapus data terapis",
                "operationId": "deleteTherapist",
                "parameters": [
                    {
                        "name": "therapistId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus"
                    },
                    "404": {
                        "description": "Data terapis tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Email Verification"
                ],
                "summary": "Verifikasi alamat email pengguna",
                "description": "Endpoint ini diakses dari link yang dikirim ke email pengguna. Memerlukan signature yang valid.",
                "operationId": "verifyEmail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email berhasil diverifikasi"
                    },
                    "400": {
                        "description": "Tautan tidak valid atau kadaluwarsa"
                    },
                    "404": {
                        "description": "Pengguna tidak ditemukan"
                    }
                }
            }
        },
        "/email/resend-verification": {
            "post": {
                "tags": [
                    "Email Verification"
                ],
                "summary": "Kirim ulang email verifikasi",
                "description": "Mengirim ulang email verifikasi ke pengguna yang sedang login.",
                "operationId": "resendVerification",
                "responses": {
                    "202": {
                        "description": "Link baru berhasil dikirim"
                    },
                    "200": {
                        "description": "Email sudah terverifikasi sebelumnya"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TherapistCreateRequest": {
                "required": [
                    "username",
                    "email",
                    "password",
                    "therapist_name",
                    "therapist_section",
                    "therapist_phone"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "dr.johndoe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "dr.johndoe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "Password123!"
                    },
                    "therapist_name": {
                        "type": "string",
                        "example": "Dr. John Doe"
                    },
                    "therapist_section": {
                        "type": "string",
                        "enum": [
                            "Okupasi",
                            "Fisio",
                            "Wicara",
                            "Paedagog"
                        ],
                        "example": "Okupasi"
                    },
                    "therapist_phone": {
                        "type": "string",
                        "example": "081234567890"
                    }
                },
                "type": "object"
            },
            "TherapistUpdateRequest": {
                "description": "Semua field bersifat opsional. Hanya field yang dikirim yang akan diupdate.",
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "dr.johndoe.updated"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "dr.johndoe.new@example.com"
                    },
                    "password": {
                        "description": "Kirim hanya jika ingin mengubah password",
                        "type": "string",
                        "format": "password",
                        "example": "NewPassword123!"
                    },
                    "therapist_name": {
                        "type": "string",
                        "example": "Dr. John Doe, S.Psi"
                    },
                    "therapist_section": {
                        "type": "string",
                        "enum": [
                            "Okupasi",
                            "Fisio",
                            "Wicara",
                            "Paedagog"
                        ],
                        "example": "Fisio"
                    },
                    "therapist_phone": {
                        "type": "string",
                        "example": "081234567891"
                    }
                },
                "type": "object"
            },
...
            "TherapistResource": {
                "properties": {
                    "id": {
                        "description": "Therapist ID",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "User ID terkait",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email terapis",
                        "type": "string",
                        "format": "email"
                    },
                    "username": {
                        "description": "Username terapis",
                        "type": "string"
                    },
                    "therapist_name": {
                        "description": "Nama lengkap terapis",
                        "type": "string"
                    },
                    "therapist_section": {
                        "description": "Bagian/spesialisasi terapis",
                        "type": "string"
                    },
                    "therapist_phone": {
                        "description": "Nomor telepon terapis",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Bearer Anda",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}